+++ CHANGELOG.md
@@ -36,6 +36,11 @@ TODO: Add an updated screenshot with some fancy VST3-only plugins to the readme
   the refresh rate of a plugin's editor GUI. The default 60 updates per second
   may be too high if your computer's cannot keep up, or if you're using a host
   that never closes the editor such as Ardour.
+- Added a [compatibility
+  option](https://github.com/robbert-vdh/yabridge#compatibility-options) to
+  disable HiDPI scaling for VST3 plugins. At the moment Wine does not have
+  proper HiDPI support, so some plugins might not scale their interfaces
+  correctly when the host tells those plugins to scale their GUIs.
 
 ### Changed